add migrations

+++ app/Http/Controllers/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use App\Models\Post;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
@@ -45,7 +46,7 @@ class CategoryController extends Controller
         ];
         $request->validate($rules,$messages);
         Category::create(['name'=>$request->name]);
-        return redirect()->route('admin.category.index')->with('toast_success', 'Create Category Successfully!');
+        return redirect()->route('admin.category.index')->with('success', 'Create Category Successfully!');
 
     }
 
@@ -81,22 +82,33 @@ class CategoryController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $rules= [
+            'name' => 'required',
+
+        ];
+        $messages=[
+            'name.required' => 'You must enter name',
+        ];
+        $request->validate($rules,$messages);
         $category=Category::where('id',$id)->first();
         $category->name = empty($request->name) ? $category->name : $request->name;
         $category->save();
-        return redirect()->route('admin.category.index')->with('toast_success', 'Update Category Successfully!');
-
+        return redirect()->route('admin.category.index')->with('success', 'Update Category Successfully!');
     }
-
     /**
      * Remove the specified resource from storage.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request,$id)
     {
+        if(Post::where('category_id',$id)->first()){
+            return redirect()->route('admin.category.index')->with('success', 'Cannot Delete This Category!');
+        }
         Category::where('id',$id)->delete();
-        return redirect()->route('admin.category.index')->with('toast_success', 'Delete Category Successfully!');
+        return redirect()->route('admin.category.index')->with('success', 'Destroy Category Successfully!');
     }
+
+
 }
